Shirley, part 4
Shirley decided that she had was finished firing the gun, for now. She reached into her purse for a snack. As she pulled out the sandwich, she saw some...thing on it.
"AHHHH! A SPIDER" She screamed, then passed out onto the the controls. The guns sprang to life again, peppering the buildings with bullets ad the vehicle drove past.